try mvn install or deploy again, if...

+++ cico_release.sh
@@ -128,15 +128,32 @@ releaseWorkspaceLoader() {
 
 releaseCheServer() {
     set -x
+    tmpmvnlog=/tmp/mvn.log.txt
     if [[ $RELEASE_CHE_PARENT = "true" ]]; then
         pushd che-parent >/dev/null
-        mvn clean install -U -Pcodenvy-release -Dgpg.passphrase=$CHE_OSS_SONATYPE_PASSPHRASE
+        rm -f $tmpmvnlog || true
+        set +e
+        mvn clean install -U -Pcodenvy-release -Dgpg.passphrase=$CHE_OSS_SONATYPE_PASSPHRASE | tee $tmpmvnlog
+        set -e
+        # try maven build again if Nexus dies
+        if grep -q -E "502 - Bad Gateway|Nexus connection problem" $tmpmvnlog; then
+            rm -f $tmpmvnlog || true
+            mvn clean install -U -Pcodenvy-release -Dgpg.passphrase=$CHE_OSS_SONATYPE_PASSPHRASE 
+        fi
 
         if [ $? -eq 0 ]; then
             echo 'Build of che-parent: Success!'
             if [[ ${DEPLOY_TO_NEXUS} == "true" ]]; then
                 echo 'Deploy che-parent artifacts to nexus'
-                mvn clean deploy -Pcodenvy-release -DcreateChecksum=true -DautoReleaseAfterClose=$AUTORELEASE_ON_NEXUS -Dgpg.passphrase=$CHE_OSS_SONATYPE_PASSPHRASE
+                rm -f $tmpmvnlog || true
+                set +e
+                mvn clean deploy -Pcodenvy-release -DcreateChecksum=true -DautoReleaseAfterClose=$AUTORELEASE_ON_NEXUS -Dgpg.passphrase=$CHE_OSS_SONATYPE_PASSPHRASE | tee $tmpmvnlog
+                set -e
+                # try maven build again if Nexus dies
+                if grep -q -E "502 - Bad Gateway|Nexus connection problem" $tmpmvnlog; then
+                    rm -f $tmpmvnlog || true
+                    mvn clean deploy -Pcodenvy-release -DcreateChecksum=true -DautoReleaseAfterClose=$AUTORELEASE_ON_NEXUS -Dgpg.passphrase=$CHE_OSS_SONATYPE_PASSPHRASE
+                fi
             else
                 echo "[WARN] No deployment to Nexus as DEPLOY_TO_NEXUS = ${DEPLOY_TO_NEXUS}"
             fi
@@ -148,13 +165,29 @@ releaseCheServer() {
     fi
 
     pushd che >/dev/null
-    mvn clean install -U -Pcodenvy-release -Dgpg.passphrase=$CHE_OSS_SONATYPE_PASSPHRASE
+    rm -f $tmpmvnlog || true
+    set +e
+    mvn clean install -U -Pcodenvy-release -Dgpg.passphrase=$CHE_OSS_SONATYPE_PASSPHRASE | tee $tmpmvnlog
+    set -e
+    # try maven build again if Nexus dies
+    if grep -q -E "502 - Bad Gateway|Nexus connection problem" $tmpmvnlog; then
+        rm -f $tmpmvnlog || true
+        mvn clean install -U -Pcodenvy-release -Dgpg.passphrase=$CHE_OSS_SONATYPE_PASSPHRASE
+    fi
 
     if [ $? -eq 0 ]; then
         echo 'Build of che-server: Success!'
         if [[ ${DEPLOY_TO_NEXUS} == "true" ]]; then
             echo 'Deploy che-server artifacts to nexus'
-            mvn clean deploy -Pcodenvy-release -DcreateChecksum=true -DautoReleaseAfterClose=$AUTORELEASE_ON_NEXUS -Dgpg.passphrase=$CHE_OSS_SONATYPE_PASSPHRASE
+            rm -f $tmpmvnlog || true
+            set +e
+            mvn clean deploy -Pcodenvy-release -DcreateChecksum=true -DautoReleaseAfterClose=$AUTORELEASE_ON_NEXUS -Dgpg.passphrase=$CHE_OSS_SONATYPE_PASSPHRASE | tee $tmpmvnlog
+            set -e
+            # try maven build again if Nexus dies
+            if grep -q -E "502 - Bad Gateway|Nexus connection problem" $tmpmvnlog; then
+                rm -f $tmpmvnlog || true
+                mvn clean deploy -Pcodenvy-release -DcreateChecksum=true -DautoReleaseAfterClose=$AUTORELEASE_ON_NEXUS -Dgpg.passphrase=$CHE_OSS_SONATYPE_PASSPHRASE
+            fi
         else
             echo "[WARN] No deployment to Nexus as DEPLOY_TO_NEXUS = ${DEPLOY_TO_NEXUS}"
         fi
